+++ src/GpibInterface/GpibInterface.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <sstream>
 #include <string>
-#include "GpibInterface.h"
+#include "GpibInterface/GpibInterface.h"
 
 
 #include <windows.h>
